fix(services):fixed missing user validation in tables

+++ src/services/address.service.ts
@@ -1,10 +1,17 @@
 import { db } from "../db/client";
 import { addresses } from "../models/schema";
-import { eq } from "drizzle-orm";
+import { eq, and } from "drizzle-orm";
+import { ROLES } from "../types";
 
-export async function getAllAddresses() {
-  return await db.select().from(addresses);
-}
+export async function getAllAddresses(role:ROLES,user_id:number) {
+   if (role === ROLES.ADMIN) {
+     return await db.select().from(addresses);
+   }
+   if (role === ROLES.USER) {
+     return await db.select().from(addresses).where(eq(addresses.user_id, user_id));
+   }
+ }
+
 
 export async function getAddressById(id: number) {
   const result = await db.select().from(addresses).where(eq(addresses.id, id));
@@ -24,12 +31,14 @@ export async function createAddress(data: {
 
 
 
-export async function deleteAddress(id: number) {
-  return await db.delete(addresses).where(eq(addresses.id, id));
+export async function deleteAddress(id: number, userId: number)
+{
+  return await db.delete(addresses).where(and(eq(addresses.id,id),eq(addresses.id,userId)));
 }
 
 export async function updateAddress(
   id: number,
+  userId:number,
   data: {
     user_id?: number;
     name?: string;
@@ -38,5 +47,6 @@ export async function updateAddress(
     is_default?: boolean;
   }
 ) {
-  return await db.update(addresses).set(data).where(eq(addresses.id, id));
+
+  return await db.update(addresses).set(data).where(and(eq(addresses.id,id),eq(addresses.id,userId)));
 }
