+++ src/services/cart.service.ts
@@ -1,9 +1,15 @@
 import { db } from "../db/client";
 import { carts } from "../models/schema";
-import { eq } from "drizzle-orm";
+import { eq, and } from "drizzle-orm";
+import { ROLES } from "../types";
 
-export async function getAllCarts() {
-  return await db.select().from(carts);
+export async function getAllCarts(role: ROLES, user_id: number) {
+  if (role === ROLES.ADMIN) {
+    return await db.select().from(carts);
+  }
+  if (role === ROLES.USER) {
+    return await db.select().from(carts).where(eq(carts.user_id, user_id));
+  }
 }
 
 export async function getCartById(id: number) {
@@ -14,10 +20,10 @@ export async function createCart(data: { user_id: number }) {
   return await db.insert(carts).values(data);
 }
 
-export async function deleteCart(id: number) {
-  return await db.delete(carts).where(eq(carts.id, id));
+export async function deleteCart(id: number, userId: number) {
+  return await db.delete(carts).where(and(eq(carts.id, id), eq(carts.user_id, userId)));
 }
 
-export async function updateCart(id: number, data: { user_id?: number }) {
-  return await db.update(carts).set(data).where(eq(carts.id, id));
+export async function updateCart(id: number,userId:number, data: { user_id?: number }) {
+  return await db.update(carts).set(data).where(and(eq(carts.id, id), eq(carts.user_id, userId)));
 }
