+++ src/services/users.service.ts
@@ -1,6 +1,6 @@
 import { db } from "../db/client";
 import { users } from "../models/schema";
-import { eq } from "drizzle-orm";
+import { eq, and } from "drizzle-orm";
 
 export async function getAllUsers() {
   return await db.select().from(users);
@@ -18,14 +18,14 @@ export async function createUser(data: {
   return await db.insert(users).values(data);
 }
 
-export async function deleteUser(id: number) {
-  return await db.delete(users).where(eq(users.id, id));
+export async function deleteUser(id: number,userId:number) {
+  return await db.delete(users).where(and(eq(users.id,id),eq(users.id,userId)));
 }
 export async function updateUser(
-  id: number,
+  id: number,userId:number,
   data: { name?: string; email?: string; password?: string }
 ) {
-  return await db.update(users).set(data).where(eq(users.id, id));
+  return await db.update(users).set(data).where(and(eq(users.id,id),eq(users.id,userId)));  
 }
 
 export async function returnUserValid(userId:number) {
