+++ src/services/wishlist.service.ts
@@ -1,9 +1,15 @@
 import { db } from "../db/client";
 import { wishlists } from "../models/schema";
-import { eq } from "drizzle-orm";
+import { eq, and } from "drizzle-orm";
+import { ROLES } from "../types";
 
-export async function getAllWishlists() {
-  return await db.select().from(wishlists);
+export async function getAllWishlists(role: ROLES, user_id: number) {
+    if (role === ROLES.ADMIN) {
+      return await db.select().from(wishlists);
+    }
+    if (role === ROLES.USER) {
+      return await db.select().from(wishlists).where(eq(wishlists.user_id, user_id));
+    }
 }
 
 export async function getWishlistById(id: number) {
@@ -14,12 +20,13 @@ export async function createWishlist(data: { product_id: number; user_id: number
   return await db.insert(wishlists).values(data);
 }
 
-export async function deleteWishlist(id: number) {
-  return await db.delete(wishlists).where(eq(wishlists.id, id));
+export async function deleteWishlist(id: number,userId:number) {
+  return await db.delete(wishlists).where(and(eq(wishlists.id,id),eq(wishlists.user_id,userId)));
 }
 export async function updateWishlist(
   id: number,
+  userId:number,
   data: { user_id?: number; product_id?: number }
 ) {
-  return await db.update(wishlists).set(data).where(eq(wishlists.id, id));
+  return await db.update(wishlists).set(data).where(and(eq(wishlists.id,id),eq(wishlists.user_id,userId)));
 }
